Simplify and error handling * Simplify some of the operation creation logic. * Show informative error log when SharedArrayBuffer can't be use
