+++ emulator/README.md
@@ -16,17 +16,6 @@ The source code for this UI is also in the `ui/` folder in the package github re
 * [Default Controls](#default-controls)
 * [API](#api)
 
-### Note: https and same origin required due to SharedArrayBuffer
-In order to provide seamless near real-time buffered audio, this emulator uses a shared array buffer to
-write to an audio player on another thread. Since this allows direct memory access, certain security
-features are required. First, you must use https. In addition, you must have the following headers set
-on your server:
-
-```
-'Cross-Origin-Opener-Policy': 'same-origin'
-'Cross-Origin-Embedder-Policy': 'require-corp'
-```
-
 The simple example shown below has been implemented in a folder in the github repo named simple-example.
 If you look at the webpack config in that folder, you will see that the dev server has been configured
 with these settings.
@@ -102,8 +91,21 @@ for file selection. For simplicity, we will add that to our event handler from a
 ```js
 // enable audio
 gameboy.apu.enableSound();
+```
+
+#### Note: https and same origin required for audio to work due to SharedArrayBuffer
+In order to provide near real-time buffered audio, this emulator uses a shared array buffer to
+write to an audio player on another thread. Since this allows direct memory access, certain security
+features are required. First, you must use https. In addition, you must have the following headers set
+on your server:
 
 ```
+'Cross-Origin-Opener-Policy': 'same-origin'
+'Cross-Origin-Embedder-Policy': 'require-corp'
+```
+
+You can see these being set in the webpack dev server in the simple example projects. You can still
+use the emulator without this, but audio will not work.
 
 ### Render to Canvas
