+++ emulator/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Gameboy Example</title>
+</head>
+<body>
+<input type="file" class="file-input" style="display: block;"/>
+<canvas width="160" height="144" style="width: 320px; image-rendering: pixelated"></canvas>
