+++ emulator/package.json
@@ -1,6 +1,6 @@
 {
   "name": "gameboy-emulator",
-  "version": "1.0.2",
+  "version": "1.1.0",
   "description": "A GameBoy emulator written with typescript",
   "main": "./dist/gameboy.js",
   "module": "./dist/gameboy.js",
@@ -22,6 +22,7 @@
   },
   "scripts": {
     "build": "webpack --config webpack.npm-package.js",
+    "serve": "webpack serve --mode=development",
     "test": "jest"
   },
   "keywords": [
