--- emulator/src/cpu/operations/arithmetic-and-logical/index.ts
@@ -1,20 +1,0 @@
-import { createAddOperations } from "./add-operations";
-import { createSubtractOperations } from "./subtract-operations";
-import { createAndOperations } from "./and-operations";
-import { createOrOperations } from "./or-operations";
-import { createXorOperations } from "./xor-operations";
-import { createCompareOperations } from "./compare-operations";
-import { createIncrementOperations } from "./increment-operations";
-import { createDecrementOperations } from "./decrement-operations";
-
-export {
-  createAddOperations,
-  createSubtractOperations,
-  createAndOperations,
-  createOrOperations,
-  createXorOperations,
-  createCompareOperations,
-  createIncrementOperations,
-  createDecrementOperations,
-}
-
