+++ emulator/src/cpu/operations/cb-operations/set-operations.ts
@@ -1,11 +1,9 @@
 import { CPU } from "@/cpu/cpu";
-import { Operation } from "@/cpu/operations/operation.model";
 import { setBit } from "@/helpers/binary-helpers";
 import { memory } from "@/memory/memory";
 import { CpuRegister } from "@/cpu/internal-registers/cpu-register";
 
-export function getSetSubOperations(cpu: CPU): Operation[] {
-  const subOperations: Operation[] = [];
+export function getSetSubOperations(cpu: CPU) {
   const { registers } = cpu;
 
   // ****************R
@@ -17,7 +15,7 @@ export function getSetSubOperations(cpu: CPU): Operation[] {
 
   cpu.registers.baseRegisters.forEach(register => {
     for (let bitPosition = 0; bitPosition < 8; bitPosition++) {
-      subOperations.push({
+      cpu.addCbOperation({
         byteDefinition: getSetBAByteDefinition(bitPosition, register.code),
         instruction: `SET ${bitPosition}, ${register.name}`,
         cycleTime: 2,
@@ -38,7 +36,7 @@ export function getSetSubOperations(cpu: CPU): Operation[] {
   }
 
   for (let bitPosition = 0; bitPosition < 8; bitPosition++) {
-    subOperations.push({
+    cpu.addCbOperation({
       byteDefinition: getSetHLByteDefinition(bitPosition),
       instruction: `SET ${bitPosition}, (HL)`,
       cycleTime: 3,
@@ -50,6 +48,4 @@ export function getSetSubOperations(cpu: CPU): Operation[] {
       }
     })
   }
-
-  return subOperations;
 }
