+++ emulator/src/cpu/operations/create-general-purpose-operations.ts
@@ -1,11 +1,9 @@
-import { Operation } from "../operation.model";
 import { CPU } from "@/cpu/cpu";
 
-export function createGeneralPurposeOperations(cpu: CPU): Operation[] {
-  const generalPurposeOperations: Operation[] = [];
-  const { registers } = cpu;
+export function createGeneralPurposeOperations(this: CPU) {
+  const { registers } = this;
 
-  generalPurposeOperations.push({
+  this.addOperation({
     instruction: 'DAA',
     byteDefinition: 0x27,
     cycleTime: 1,
@@ -32,7 +30,7 @@ export function createGeneralPurposeOperations(cpu: CPU): Operation[] {
     }
   });
 
-  generalPurposeOperations.push({
+  this.addOperation({
     instruction: 'CPL',
     byteDefinition: 0b00_101_111,
     cycleTime: 1,
@@ -44,7 +42,7 @@ export function createGeneralPurposeOperations(cpu: CPU): Operation[] {
     }
   })
 
-  generalPurposeOperations.push({
+  this.addOperation({
     instruction: 'NOP',
     byteDefinition: 0b0,
     cycleTime: 1,
@@ -54,7 +52,7 @@ export function createGeneralPurposeOperations(cpu: CPU): Operation[] {
     }
   });
 
-  generalPurposeOperations.push({
+  this.addOperation({
     instruction: 'CCF',
     byteDefinition: 0x3f,
     cycleTime: 1,
@@ -67,7 +65,7 @@ export function createGeneralPurposeOperations(cpu: CPU): Operation[] {
     }
   });
 
-  generalPurposeOperations.push({
+  this.addOperation({
     instruction: 'SCF',
     byteDefinition: 0x37,
     cycleTime: 1,
@@ -80,25 +78,23 @@ export function createGeneralPurposeOperations(cpu: CPU): Operation[] {
     }
   });
 
-  generalPurposeOperations.push({
+  this.addOperation({
     instruction: 'HALT',
     byteDefinition: 0x76,
     cycleTime: 1,
     byteLength: 1,
-    execute() {
-      cpu.halt();
+    execute: () => {
+      this.halt();
     }
   });
 
-  generalPurposeOperations.push({
+  this.addOperation({
     instruction: 'STOP',
     byteDefinition: 0x10,
     cycleTime: 1,
     byteLength: 1,
-    execute() {
-      cpu.stop();
+    execute: () => {
+      this.stop();
     }
   });
-
-  return generalPurposeOperations;
 }
